Add PhonesBlocConsumer Widget for Home View
This commit introduces the `PhonesBlocConsumer` widget, which is responsible for consuming the `PhonesDataCubit` state and rendering the appropriate UI based on the state. The widget listens to the `PhonesDataCubit` and displays the `HomeViewBody` when the state is `PhonesDataSuccess`, a `CustomErrorWidget` with the error message when the state is `PhonesDataFailure`, and a `CustomProgressIndicator` for all other states. This ensures a smooth and responsive user experience by handling different states of the data fetching process.