Refactor HomeView to use BlocProvider and PhonesDataCubit
Updated `HomeView` to use `BlocProvider` for managing the state of `PhonesDataCubit`. The `HomeViewBody` widget now receives a list of `PhoneEntites` as a parameter, which is passed from the `PhonesBlocConsumer` when the state is `PhonesDataSuccess`. This change ensures that the home view dynamically fetches and displays phone data using the `PhonesDataCubit` and `HomeRepo`. Additionally, the `HomeViewBody` widget has been updated to accept the list of phones, making it more flexible and reusable.