Fix phone type and name mapping in AdminRepoImpl
Corrected the mapping of phone type and name in the `AdminRepoImpl` class. Previously, the `type` was being set to the value of `data["phoneType"]` and the `name` was being set to `data["phoneName"]`. This commit swaps these values so that the `type` is now set to the uppercase version of `data["phoneName"]` and the `name` is set to `data["phoneType"]`. This change ensures that the phone type and name are correctly mapped according to the intended logic.